+++ platform/viewer/src/connectedComponents/ViewerRetrieveStudyData.js
@@ -1,11 +1,11 @@
-import React, { Component } from "react";
-import { metadata, studies, utils } from "@ohif/core";
+import React, { Component } from 'react';
+import { metadata, studies, utils, log } from '@ohif/core';
 
-import ConnectedViewer from "./ConnectedViewer.js";
-import PropTypes from "prop-types";
-import { extensionManager } from "./../App.js";
+import ConnectedViewer from './ConnectedViewer.js';
+import PropTypes from 'prop-types';
+import { extensionManager } from './../App.js';
 
-const { OHIFStudyMetadata } = metadata;
+const { OHIFStudyMetadata, OHIFSeriesMetadata } = metadata;
 const { retrieveStudiesMetadata } = studies;
 const { studyMetadataManager, updateMetaDataManager } = utils;
 
@@ -13,63 +13,95 @@ class ViewerRetrieveStudyData extends Component {
   static propTypes = {
     studyInstanceUids: PropTypes.array.isRequired,
     seriesInstanceUids: PropTypes.array,
-    server: PropTypes.object
+    server: PropTypes.object,
   };
 
-  state = {
-    studies: null,
-    error: null
-  };
+  constructor(props) {
+    super(props);
+    this.state = {
+      studies: null,
+      error: null,
+    };
+  }
 
-  async componentDidMount() {
-    // TODO: Avoid using timepoints here
-    //const params = { studyInstanceUids, seriesInstanceUids, timepointId, timepointsFilter={} };
-    const { studyInstanceUids, seriesInstanceUids, server } = this.props;
-
+  async loadStudies() {
     try {
+      const { server, studyInstanceUids, seriesInstanceUids } = this.props;
       const studies = await retrieveStudiesMetadata(
         server,
         studyInstanceUids,
         seriesInstanceUids
       );
+      this.setStudies(studies);
+    } catch (e) {
+      this.setState({ error: true });
+      log.error(e);
+    }
+  }
 
-      // Render the viewer when the data is ready
-      // TODO: CLEAR THIS SOMEWHERE ELSE
-      studyMetadataManager.purge();
-
+  setStudies(givenStudies) {
+    if (Array.isArray(givenStudies) && givenStudies.length > 0) {
+      const sopClassHandlerModules =
+        extensionManager.modules['sopClassHandlerModule'];
       // Map studies to new format, update metadata manager?
-      const updatedStudies = studies.map(study => {
+      const studies = givenStudies.map(study => {
         const studyMetadata = new OHIFStudyMetadata(
           study,
           study.studyInstanceUid
         );
-        const sopClassHandlerModules =
-          extensionManager.modules["sopClassHandlerModule"];
-
-        study.displaySets =
-          study.displaySets ||
-          studyMetadata.createDisplaySets(sopClassHandlerModules);
+        if (!study.displaySets) {
+          study.displaySets = studyMetadata.createDisplaySets(
+            sopClassHandlerModules
+          );
+        }
         studyMetadata.setDisplaySets(study.displaySets);
-
         // Updates WADO-RS metaDataManager
         updateMetaDataManager(study);
-
         studyMetadataManager.add(studyMetadata);
-
+        // Attempt to load remaning series if any
+        this._attemptToLoadRemainingSeries(studyMetadata);
         return study;
       });
-
-      this.setState({
-        studies: updatedStudies
-      });
-    } catch (err) {
-      this.setState({
-        error: true
-      });
-
-      // TODO: Handle gracefully instead of throwing?
-      throw new Error(err);
+      this.setState({ studies });
     }
+  }
+
+  _addSeriesToStudy(studyMetadata, series) {
+    const sopClassHandlerModules =
+      extensionManager.modules['sopClassHandlerModule'];
+    const study = studyMetadata.getData();
+    const seriesMetadata = new OHIFSeriesMetadata(series, study);
+    studyMetadata.addSeries(seriesMetadata);
+    studyMetadata.createAndAddDisplaySetsForSeries(
+      sopClassHandlerModules,
+      seriesMetadata
+    );
+    study.displaySets = studyMetadata.getDisplaySets();
+    updateMetaDataManager(study, series.seriesInstanceUid);
+    this.setState(function(state) {
+      return { studies: state.studies.slice() };
+    });
+  }
+
+  _attemptToLoadRemainingSeries(studyMetadata) {
+    const { seriesLoader } = studyMetadata.getData();
+    if (!seriesLoader) {
+      return;
+    }
+    while (seriesLoader.hasNext()) {
+      seriesLoader
+        .next()
+        .then(
+          series => void this._addSeriesToStudy(studyMetadata, series),
+          error => void log.error(error)
+        );
+    }
+  }
+
+  componentDidMount() {
+    // TODO: CLEAR THIS SOMEWHERE ELSE
+    studyMetadataManager.purge();
+    this.loadStudies();
   }
 
   render() {